refactor: Consolidate different activities in one place
* Move screens
