+++ common/src/main/java/com/wynntils/screens/activities/WynntilsDiscoveriesScreen.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.screens.discoveries;
+package com.wynntils.screens.activities;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.wynntils.core.WynntilsMod;
@@ -11,6 +11,8 @@ import com.wynntils.core.text.StyledText;
 import com.wynntils.models.activities.event.ActivityUpdatedEvent;
 import com.wynntils.models.activities.type.ActivitySortOrder;
 import com.wynntils.models.discoveries.DiscoveryInfo;
+import com.wynntils.screens.activities.widgets.DiscoveryButton;
+import com.wynntils.screens.activities.widgets.DiscoveryProgressButton;
 import com.wynntils.screens.base.WynntilsListScreen;
 import com.wynntils.screens.base.widgets.BackButton;
 import com.wynntils.screens.base.widgets.FilterButton;
@@ -18,8 +20,6 @@ import com.wynntils.screens.base.widgets.PageSelectorButton;
 import com.wynntils.screens.base.widgets.ReloadButton;
 import com.wynntils.screens.base.widgets.SortOrderWidget;
 import com.wynntils.screens.base.widgets.SortableActivityScreen;
-import com.wynntils.screens.discoveries.widgets.DiscoveryButton;
-import com.wynntils.screens.discoveries.widgets.DiscoveryProgressButton;
 import com.wynntils.screens.wynntilsmenu.WynntilsMenuScreen;
 import com.wynntils.utils.StringUtils;
 import com.wynntils.utils.colors.CommonColors;
