+++ common/src/main/java/com/wynntils/screens/wynntilsmenu/WynntilsMenuScreen.java
@@ -10,15 +10,15 @@ import com.wynntils.core.components.Models;
 import com.wynntils.core.text.StyledText;
 import com.wynntils.features.map.MapFeature;
 import com.wynntils.screens.activities.WynntilsCaveScreen;
+import com.wynntils.screens.activities.WynntilsDialogueHistoryScreen;
+import com.wynntils.screens.activities.WynntilsDiscoveriesScreen;
+import com.wynntils.screens.activities.WynntilsQuestBookScreen;
 import com.wynntils.screens.base.WynntilsMenuScreenBase;
-import com.wynntils.screens.discoveries.WynntilsDiscoveriesScreen;
 import com.wynntils.screens.guides.WynntilsGuidesListScreen;
 import com.wynntils.screens.lootrun.WynntilsLootrunsScreen;
 import com.wynntils.screens.maps.MainMapScreen;
 import com.wynntils.screens.maps.PoiManagementScreen;
 import com.wynntils.screens.overlays.selection.OverlaySelectionScreen;
-import com.wynntils.screens.questbook.WynntilsQuestBookScreen;
-import com.wynntils.screens.questbook.history.WynntilsDialogueHistoryScreen;
 import com.wynntils.screens.settings.WynntilsBookSettingsScreen;
 import com.wynntils.screens.wynntilsmenu.widgets.WynntilsMenuButton;
 import com.wynntils.utils.colors.CommonColors;
